Move  to BaseDataset.
